refactor: split matching logic to match.go file and clean up codebase
